remove TaskInProjectDataSource interface
