+++ src/main/kotlin/data/source/TaskDataSource.kt
@@ -2,6 +2,7 @@ package org.example.data.source
 
 import data.dto.MateTaskAssignmentDto
 import data.dto.TaskDto
+import data.dto.TaskInProjectDto
 
 interface TaskDataSource {
     suspend fun getAllTasks(): List<TaskDto>
@@ -10,10 +11,13 @@ interface TaskDataSource {
     suspend fun deleteTask(id: String): Boolean
     suspend fun getTasksInProject(projectId: String): List<TaskDto>
     suspend fun getTasksByIds(taskIds: List<String>): List<TaskDto>
-    suspend fun addTaskInProject(projectId: String, taskId: String): Boolean
-    suspend fun deleteTaskFromProject(projectId: String, taskId: String): Boolean
     suspend fun getAllTasksByUserName(username: String): List<TaskDto>
 
     suspend fun getUsersMateTaskByTaskId(taskId: String): List<MateTaskAssignmentDto>
     suspend fun getUsersMateTaskByUserName(username: String): List<MateTaskAssignmentDto>
+
+    suspend fun addTaskInProject(projectId: String, taskId: String): Boolean
+    suspend fun deleteTaskFromProject(projectId: String, taskId: String): Boolean
+    suspend fun getAllTasksInProject(): List<TaskInProjectDto>
+    suspend fun getTasksInProjectByProjectId(projectId: String): List<TaskInProjectDto>
 }
