+++ src/main/kotlin/data/source/remote/mongo/TaskMongoDataSource.kt
@@ -1,19 +1,20 @@
 package data.source.remote.mongo
 
+import com.mongodb.client.model.Filters
 import com.mongodb.kotlin.client.coroutine.MongoCollection
 import data.dto.MateTaskAssignmentDto
 import data.dto.TaskDto
+import data.dto.TaskInProjectDto
 import kotlinx.coroutines.flow.toList
 import org.example.data.source.TaskDataSource
-import org.example.data.source.TaskInProjectDataSource
-import org.litote.kmongo.and
+
 import org.litote.kmongo.eq
 import org.litote.kmongo.`in`
 
 class TaskMongoDataSource(
     private val taskCollection: MongoCollection<TaskDto>,
     private val mateTaskAssignmentCollection: MongoCollection<MateTaskAssignmentDto>,
-    private val taskInProjectDataSource: TaskInProjectDataSource,
+    private val taskInProjectCollection: MongoCollection<TaskInProjectDto>,
 ) : TaskDataSource {
 
     override suspend fun getAllTasks(): List<TaskDto> = taskCollection.find().toList()
@@ -27,18 +28,12 @@ class TaskMongoDataSource(
         taskCollection.deleteOne(TaskDto::id eq id).deletedCount > 0
 
     override suspend fun getTasksInProject(projectId: String): List<TaskDto> =
-        taskInProjectDataSource.getTasksInProjectByProjectId(projectId).map { it.taskId }.let { taskIds ->
+        getTasksInProjectByProjectId(projectId).map { it.taskId }.let { taskIds ->
             getTasksByIds(taskIds)
         }
 
     override suspend fun getTasksByIds(taskIds: List<String>): List<TaskDto> =
         taskCollection.find(TaskDto::id `in` taskIds).toList()
-
-    override suspend fun addTaskInProject(projectId: String, taskId: String): Boolean =
-        taskInProjectDataSource.addTaskInProject(projectId, taskId)
-
-    override suspend fun deleteTaskFromProject(projectId: String, taskId: String): Boolean =
-        taskInProjectDataSource.deleteTaskFromProject(projectId = projectId, taskId = taskId)
 
     override suspend fun getAllTasksByUserName(username: String): List<TaskDto> =
         getUsersMateTaskByUserName(username).map { it.taskId }
@@ -49,4 +44,21 @@ class TaskMongoDataSource(
 
     override suspend fun getUsersMateTaskByUserName(username: String): List<MateTaskAssignmentDto> =
         mateTaskAssignmentCollection.find(MateTaskAssignmentDto::username eq username).toList()
+
+    override suspend fun addTaskInProject(projectId: String, taskId: String): Boolean =
+        taskInProjectCollection.insertOne(TaskInProjectDto(taskId, projectId)).insertedId != null
+
+    override suspend fun deleteTaskFromProject(projectId: String, taskId: String): Boolean =
+        taskInProjectCollection.deleteOne(
+            Filters.and(
+                Filters.eq(TaskInProjectDto::projectId.name, projectId),
+                Filters.eq(TaskInProjectDto::taskId.name, taskId)
+            )
+        ).deletedCount > 0
+
+    override suspend fun getAllTasksInProject(): List<TaskInProjectDto> =
+        taskInProjectCollection.find().toList()
+
+    override suspend fun getTasksInProjectByProjectId(projectId: String): List<TaskInProjectDto> =
+        taskInProjectCollection.find(TaskInProjectDto::projectId eq projectId).toList()
 }
